Add files via upload
This project identifies which out of the NYC schools has the best math results, the top 10 performing schools based on the combined SAT scores and which single borough has the largest standard deviation in the combined SAT score. The project uses data manipulation and summary statistics to analyse test scores, which involves extracting and transforming data.