Add files via upload
This projects performs an exploratory data analysis on the netflix_data.csv data to understand more about movies from the 1990s decade. It identifies the most frequent movie duration in the 1990s and the number of short (less than 90 minutes)action movies released in the 1990s. This project involves applying foundational Python skills to manipulate and visualize movie data